Polymarket-scanner/requirements.txt
# Core dependencies
aiohttp>=3.9.1
discord.py>=2.3.2
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Data processing
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Machine learning (for calibration)
scikit-learn>=1.3.2

# Optional: for better async performance
uvloop==0.19.0; sys_platform != 'win32'

# Optional: for database storage
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23
